second round of code reviews
