+++ .env
@@ -1,2 +1,4 @@
-GET_TAGS_URL='https://unfinished-api.herokuapp.com/api/tags'
-GET_PROJECTS_URL='https://unfinished-api.herokuapp.com/api/projects'
+REACT_APP_TAGS_URL=https://unfinished-api.herokuapp.com/api/tags
+REACT_APP_PROJECTS_URL=https://unfinished-api.herokuapp.com/api/projects
+PORT=3500
+REACT_APP_API_KEY=abcde
