+++ package.json
@@ -1,11 +1,12 @@
 {
-  "name": "capstone-project",
+  "name": "unfinished",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "dotenv": "^8.2.0",
     "fuzzysearch": "^1.0.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
